JKASFJHjksahdfash

+++ tp3Orga2/entregable/tp3-bundle.v1/src/screen.h
@@ -22,9 +22,9 @@ typedef struct ca_s {
     uint8_t a;
 } ca;
 
-void print(const char* text, uint32_t x, uint32_t y, uint16_t attr);
-void print_dec(uint32_t numero, uint32_t size, uint32_t x, uint32_t y, uint16_t attr);
-void print_hex(uint32_t numero, int32_t size, uint32_t x, uint32_t y, uint16_t attr);
+void print(const char* text, uint32_t x, uint32_t y, uint8_t attr);
+void print_dec(uint32_t numero, uint32_t size, uint32_t x, uint32_t y, uint8_t attr);
+void print_hex(uint32_t numero, int32_t size, uint32_t x, uint32_t y, uint8_t attr);
 
 void screen_drawBox(uint32_t fInit, uint32_t cInit,
                     uint32_t fSize, uint32_t cSize,
